requirements.txt
backports.functools-lru-cache==1.6.1
click==7.1.1
Flask==1.1.1
gunicorn==19.10.0
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
MarkupSafe==1.1.1
nltk==3.5
numpy==1.16.6
regex==2020.4.4
soupsieve==1.9.5
tqdm==4.45.0
Werkzeug==1.0.0
requests==2.22.0
webvtt-py
pycaption
